+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Changed
+
+- Made remote follow alert translatable.
+
 ## [4.13.0] - 2025-11-13
 
 ### Added