+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Made remote follow alert translatable.
 - Reduced space between words in post timestamp.
 - Updated translations ([#142](https://codeberg.org/silverpill/mitra-web/pulls/142)).
+- Change color of "replying to" element on hover.
 
 ### Fixed