+++ CHANGELOG.md
@@ -22,6 +22,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Prevent display name in notification header from shrinking too much on small screens.
+- Prevent overflow in poll options.
 
 ## [4.13.0] - 2025-11-13