+++ CHANGELOG.md
@@ -18,7 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Made remote follow alert translatable.
 - Reduced space between words in post timestamp.
-- Updated translations ([#142](https://codeberg.org/silverpill/mitra-web/pulls/142)).
+- Updated translations ([#142](https://codeberg.org/silverpill/mitra-web/pulls/142), [#143](https://codeberg.org/silverpill/mitra-web/pulls/143)).
 - Change color of "replying to" element on hover.
 - Increased line height in poll footer.
 - Register client before logging in.