+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [4.14.0] - 2025-11-28
+
 ### Added
 
 - Suggest replying from the home server when guest clicks on a reply button.